[IMP] delivery_carrier_label_batch: tracking reference purge
Sometimes, when regenerating labels from third party API, we could not get the
new tracking references. And then as we do not purge tracking references
correctly, previous ones are kept and lead to incoherences (batch picking
tracking references emptied but labels still existing). For this we need to
manually purge references when asking for full regeneration.