Fix dc_package_measure_required for not done transfer
The required dimensions on a package need not to be enforce before the
picking is being done (package with content).
Because that blocks any preparation on the package before it is
delivered.

But we still want to be able to check if the package is valid at will.
This can be done with the `delivery_pkg_measure__ignore_package_content`
context key.